re-order files and modules to compile in async and no-std
