+++ bin/tx-prover/src/api/mod.rs
@@ -3,14 +3,13 @@ use miden_tx::{
     utils::{Deserializable, Serializable},
     LocalTransactionProver, TransactionProver,
 };
-use tokio::{net::TcpListener, sync::Mutex};
-use tonic::{Request, Response, Status};
-use tracing::info;
-
-use crate::generated::{
+use miden_tx_prover::generated::{
     api_server::{Api as ProverApi, ApiServer},
     ProveTransactionRequest, ProveTransactionResponse,
 };
+use tokio::{net::TcpListener, sync::Mutex};
+use tonic::{Request, Response, Status};
+use tracing::info;
 
 pub struct RpcListener {
     pub api_service: ApiServer<ProverRpcApi>,
